requirements.txt
Django==1.9.13
requests==2.17.3
sanic==0.6.0
pytest-asyncio==0.8.0
asynctest==0.10.1
selenium==3.6.0
aiohttp==2.2.5
pyyaml==3.12
git+https://github.com/zero-88/hypersh_client.git@v0.1.0#egg=hypersh-client-0.1.0